requirements.txt
Django==2.2.2
djangorestframework==3.9.4
mutagen==1.42.0
pytz==2019.1
sqlparse==0.3.0
